requirements.txt
paho-mqtt
slackclient
